GenerateIntegratedReference

This function uses cFIT to create an integrated reference for the provided seurat object

Usage

GenerateIntegratedReference(
  seuratObj,
  batchIdCol = NULL,
  targetBatchSize = 5000,
  assay = "RNA",
  ngenes = 3000,
  saveFile = NULL,
  seed = 0,
  verbose = FALSE,
  minBatchSize = 100,
  selectVariableFeaturesFromWholeInput = TRUE
)

Arguments

seuratObj

The seurat object

batchIdCol

If provided, data will be subset using this metadata field

targetBatchSize

This is an alternative to batchIdCol. If batchIdCol is not provided, data will be randomly subset

assay

The seurat assay holding count data

ngenes

The number of variable genes to select

saveFile

If provided, the integrated reference object will be saved to this file as RDS

seed

The random seed

verbose

Determined output verbosity

minBatchSize

If any resulting batch has fewer than this many cells, an error is thrown

selectVariableFeaturesFromWholeInput

If true, the variable genes will be calculated from the intact input object. If false, FindVariableFeatures() will be calculated independently on each subset seurat object, followed by Seurat::SelectIntegrationFeatures